Reverting changes in other compilers
